Committing to get some support on duplicate key error.

+++ d2/models/hero.py
@@ -2,7 +2,7 @@ from d2.models.base import Base
 from d2.models.guide  import GuideModel
 from d2.models.skill  import SkillModel
 from sqlalchemy import ForeignKey
-from sqlalchemy import Column, Integer, String, Date, DateTime
+from sqlalchemy import Column, Integer, String, Date, DateTime, Float
 from sqlalchemy.orm import relationship
 
 class HeroModel(Base):
@@ -16,26 +16,26 @@ class HeroModel(Base):
     faction = Column(String(10))
     stat = Column(String(15))
     roles = Column(String(50))
-    strength = Column(Integer)
-    agility = Column(Integer)
-    intelligence = Column(Integer)
-    strength_gain = Column(Integer)
-    agility_gain = Column(Integer)
-    intelligence_gain = Column(Integer)
+    strength = Column(Float)
+    agility = Column(Float)
+    intelligence = Column(Float)
+    strength_gain = Column(Float)
+    agility_gain = Column(Float)
+    intelligence_gain = Column(Float)
     min_hp = Column(Integer)
     max_hp = Column(Integer)
     min_mana = Column(Integer)
     max_mana = Column(Integer)
     min_damage = Column(Integer)
     max_damage = Column(Integer)
-    armor = Column(Integer)
+    armor = Column(Float)
     movespeed = Column(Integer)
     attack_range = Column(Integer)
-    min_attack_animation = Column(Integer)
-    max_attack_animation = Column(Integer)
-    min_cast_animation = Column(Integer)
-    max_cast_animation = Column(Integer)
-    base_attack_time = Column(Integer)
+    min_attack_animation = Column(Float)
+    max_attack_animation = Column(Float)
+    min_cast_animation = Column(Float)
+    max_cast_animation = Column(Float)
+    base_attack_time = Column(Float)
     missile_speed = Column(Integer)
     day_site_range = Column(Integer)
     night_site_range = Column(Integer)
